+++ back/src/api/auth.ts
@@ -1,18 +1,12 @@
-
+import express from "express";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcryptjs";
 import config from "../config";
+import auth from "../middleware/auth";
 import { check, validationResult } from "express-validator";
-
+import User from "../models/User";
 
 const router = express.Router();
-
-import User from "../models/User";
-import auth from "../middleware/auth";
-import Profile from "../models/Profile";
-
-import express, {Request, Response} from "express";
-import { IProfileInputDTO } from "../interfaces/IProfile";
 
 /**
  *  @route Post api/auth
@@ -71,14 +65,14 @@ router.post(
   }
 );
 
-/*
+/**
  *  @route GET api/auth
  *  @desc Test Route
  *  @access Public
  */
 router.get("/", auth, async function (req, res) {
-  try {
-    console.log(req.body.user);
+  try { // 매개변수 auth 라는 middleware를 통과해야 try문 돌릴 수 있음
+      console.log(req.body.user);
     const user = await User.findById(req.body.user.id).select("-password");
     res.json(user);
   } catch (err) {
@@ -86,108 +80,5 @@ router.get("/", auth, async function (req, res) {
     res.status(500).send("Server Err");
   }
 });
-
-/**
- *  @route GET api/profile/me
- *  @desc Get current users profile
- *  @access Private
- */
-router.get("/me", auth, async (req, res) => {
-  try {
-    const profile = await Profile.findOne({
-      user: req.body.user.id,
-    }).populate("user", ["name", "avatar"]);
-
-    if (!profile) {
-      return res.status(400).json({ msg: "There is no profile for this user" });
-    }
-    res.json(profile);
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send("Server Error");
-  }
-});
-
-
-/**
- *  @route POST api/profile
- *  @desc Create or update user profile
- *  @access Private
- */
-router.post(
-  "/",
-  auth,
-  [
-    check("status", "Status is required").not().isEmpty(),
-    check("skills", "Skills is required").not().isEmpty(),
-  ],
-  async (req: Request, res: Response) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const {
-      company,
-      website,
-      location,
-      bio,
-      status,
-      githubusername,
-      skills,
-      youtube,
-      facebook,
-      twitter,
-      linkedin,
-      instagram,
-      user,
-    } = req.body;
-
-    // Build profile object
-    let profileFields: IProfileInputDTO = {
-      user: user.id,
-    };
-    if (company) profileFields.company = company;
-    if (website) profileFields.website = website;
-    if (location) profileFields.location = location;
-    if (bio) profileFields.bio = bio;
-    if (status) profileFields.status = status;
-    if (githubusername) profileFields.githubusername = githubusername;
-    if (skills) {
-      profileFields.skills = skills.split(",").map((skill) => skill.trim());
-    }
-
-    // Build social object
-
-    if (youtube) profileFields.social.youtube = youtube;
-    if (facebook) profileFields.social.facebook = facebook;
-    if (twitter) profileFields.social.twitter = twitter;
-    if (linkedin) profileFields.social.linkedin = linkedin;
-    if (instagram) profileFields.social.instagram = instagram;
-
-    try {
-      let profile = await Profile.findOne({ user: user.id });
-
-      if (profile) {
-        // Update
-        profile = await Profile.findOneAndUpdate(
-          { user: user.id },
-          { $set: { value: profileFields } },
-          { new: true }
-        );
-
-        return res.json(profile);
-      }
-
-      // Create
-      profile = new Profile(profileFields);
-      await profile.save();
-      res.json(profile);
-    } catch (err) {
-      console.error(err.message);
-      status(500).send("Server Error.");
-    }
-  }
-);
 
 module.exports = router;
