+++ back/src/interfaces/IProfileInputDTO.ts
@@ -1,22 +1,7 @@
 import mongoose from "mongoose";
-import { IExperience } from "./IExperience";
 import { IEducation } from "./IEducation";
+import { IExperience } from "./IExperience";
 import { ISocial } from "./ISocial";
-
-export interface IProfile {
-  user: mongoose.Types.ObjectId;
-  company: string;
-  website: string;
-  location: string;
-  status: string;
-  skills: [string];
-  bio: string;
-  githubusername: string;
-  experience: [IExperience];
-  education: [IEducation];
-  social: [ISocial];
-  date: Date;
-}
 
 export interface IProfileInputDTO {
   user: mongoose.Types.ObjectId;
@@ -29,6 +14,5 @@ export interface IProfileInputDTO {
   githubusername?: string;
   experience?: [IExperience];
   education?: [IEducation];
-  social?: [ISocial];
-  date?: Date;
+  social?: ISocial;
 }
