+++ back/src/interfaces/ISocial.ts
@@ -1,7 +1,7 @@
 export interface ISocial {
   youtube: String;
-  twitter: String;
-  facebook: String;
-  linkedin: String;
-  instagram: String;
+  twitter?: String;
+  facebook?: String;
+  linkedin?: String;
+  instagram?: String;
 }
